task/requirements.txt
celery==5.0.5
python-dotenv==0.14.0
